refactor: move utilsWeb3.ts to curve-lib
