+++ apps/main/src/components/PageRouterSwap/utils.ts
@@ -8,7 +8,7 @@ import uniq from 'lodash/uniq'
 
 import { NETWORK_TOKEN } from '@/constants'
 import { log } from '@/utils'
-import { weiToEther } from '@/ui/utils/utilsWeb3'
+import { weiToEther } from '@/shared/curve-lib'
 
 export const DEFAULT_FORM_STATUS: FormStatus = {
   isApproved: false,
