+++ packages/ui/src/utils/index.ts
@@ -4,7 +4,6 @@ export * from './sharedStyles'
 export * from './utilsGas'
 export * from './utilsLocale'
 export * from './utilsFormat'
-export * from './utilsWeb3'
 export * from './utilsConnectState'
 export * from './utilsNetworks'
 export * from './utilsConstants'
